Implement "status: 'deleted'"
This is only half the implementation, because the clients where the
deletion did not happen retain the note as live. The next synchronization
then upgrades the status to 'deleted', but the items remain in the
local storage.

Items with the status 'deleted' should be deleted on the client when
receiving that status from the mothership.

This means that for display, the client should filter deleted items
just like the server does (duh)