Actually invoke $EDITOR for editing
Currently the edit format is YAML front matter style, like many
blog generators use. This uses YAML for the headers and (HTML) text
for the rest. This is good enough for now.

No transition to Markdown is planned as that would mean I would need to
convert between HTML on the web client and Markdown for the command line...